Fix keepalived failover/failback behavior w/auto
Reworked the AUTO state code to properly handle failing when haproxy is
unavailable and to not fail back when it a failed peer comes back.  The
latter is to ensure we don't disrupt existing sessions.

Added a no_weight parameter that defaults to true.  Unless this is
changed, then the 'weight' parameter on the script won't be specified.
This is because if the weight isn't provided and the script fails, then
the VRRP instance will fault.